+++ examples/axum/src/graphql/input/review_input.rs
@@ -4,6 +4,6 @@ use rusty_gql::*;
 
 #[derive(GqlInputObject)]
 pub struct ReviewInput {
-    pub stars: i64,
+    pub stars: i32,
     pub commentary: Option<String>,
 }
